Gerado Model View e Controller do Cart

+++ Shop/Controllers/AddressesController.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
@@ -15,6 +16,7 @@ namespace Shop.Controllers
     {
         private readonly ShopContext _context;
         private readonly UserService _userService;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
         public AddressesController(ShopContext context, UserService userService)
         {
@@ -22,15 +24,58 @@ namespace Shop.Controllers
             _userService = userService;
         }
 
+
+        public async Task<IActionResult> MyAddress()
+        {
+            if (_userService.VerificarAutenticacao() == 0)
+            {
+                return RedirectToAction("Login", "Users");
+            }
+            int idUser = 0;
+            try
+            {
+                 idUser = Convert.ToInt32(HttpContext.Session.GetString("IdUsuario"));
+            }
+            catch
+            {
+
+            }
+
+
+            return View(await _context.Address.Where(x => x.UserId == idUser).ToListAsync());
+            
+        }
+
+        public async Task<IActionResult> AddressUser()
+        {
+            if (_userService.VerificarAutenticacao() == 0)
+            {
+                return RedirectToAction("Login", "Users");
+            }
+
+            return View();
+        }
+
+
         // GET: Addresses
         public async Task<IActionResult> Index()
         {
+            if (_userService.VerificarAutenticacao() == 0 || _userService.VerificarAutenticacao() == 1)
+            {
+                return RedirectToAction("Login", "Users");
+            }
+
             return View(await _context.Address.ToListAsync());
         }
 
         // GET: Addresses/Details/5
         public async Task<IActionResult> Details(int? id)
         {
+            if (_userService.VerificarAutenticacao() == 0)
+            {
+                return RedirectToAction("Login", "Users");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -49,6 +94,11 @@ namespace Shop.Controllers
         // GET: Addresses/Create
         public async Task<IActionResult> Create()
         {
+            if (_userService.VerificarAutenticacao() == 0 || _userService.VerificarAutenticacao() == 1)
+            {
+                return RedirectToAction("Login", "Users");
+            }
+
             var users = await _userService.FindAll();
             var viewModel = new AddressViewModel { User = users };
             return View(viewModel);
@@ -61,18 +111,52 @@ namespace Shop.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Create(/*[Bind("Id,State,City,Neighborhood,Street,Number,Telephone")]*/ Address address)
         {
-            if (ModelState.IsValid)
+            if (_userService.VerificarAutenticacao() == 0)
             {
-                _context.Add(address);
-                await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Login", "Users");
             }
+
+            try { address.UserId = Convert.ToInt32(HttpContext.Session.GetString("IdUsuario")); } catch { }
+
+            try
+            {
+                if (HttpContext.Session.GetString("UserGroupId") == "2")
+                {
+                    if (ModelState.IsValid)
+                    {
+                        _context.Add(address);
+                        await _context.SaveChangesAsync();
+                        return RedirectToAction(nameof(Index));
+                    }
+                }
+                if (HttpContext.Session.GetString("UserGroupId") == "1")
+                {
+                    if (ModelState.IsValid)
+                    {
+                        _context.Add(address);
+                        await _context.SaveChangesAsync();
+                        return RedirectToAction("MyAddress", "Addresses");
+                    }
+                }
+            }
+            catch
+            {
+
+            }
+            
+            
+            
             return View(address);
         }
 
         // GET: Addresses/Edit/5
         public async Task<IActionResult> Edit(int? id)
         {
+            if (_userService.VerificarAutenticacao() == 0 || _userService.VerificarAutenticacao() == 1)
+            {
+                return RedirectToAction("Login", "Users");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -93,6 +177,11 @@ namespace Shop.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> Edit(int id, [Bind("Id,State,City,Neighborhood,Street,Number,Telephone")] Address address)
         {
+            if (_userService.VerificarAutenticacao() == 0 || _userService.VerificarAutenticacao() == 1)
+            {
+                return RedirectToAction("Login", "Users");
+            }
+
             if (id != address.Id)
             {
                 return NotFound();
@@ -124,6 +213,11 @@ namespace Shop.Controllers
         // GET: Addresses/Delete/5
         public async Task<IActionResult> Delete(int? id)
         {
+            if (_userService.VerificarAutenticacao() == 0 )
+            {
+                return RedirectToAction("Login", "Users");
+            }
+
             if (id == null)
             {
                 return NotFound();
@@ -144,6 +238,11 @@ namespace Shop.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
+            if (_userService.VerificarAutenticacao() == 0 )
+            {
+                return RedirectToAction("Login", "Users");
+            }
+
             var address = await _context.Address.FindAsync(id);
             _context.Address.Remove(address);
             await _context.SaveChangesAsync();
